+ repaired font data structure within *h files + new 2 fonts (default and small)

+++ trunk/src/btrace.cpp
@@ -43,7 +43,7 @@ public:
         fprintf( file, "Debug Log - Error Tracer\n\n" );
         fprintf( file, "+Start Tracing\n" );
         int tab = 1;
-        for( int i=list.size()-1; i>=0; --i ) {
+        for( size_t i=list.size()-1; i>=0; --i ) {
             fprintf( file, "+" );
             for(int k=0; k<tab; ++k) fprintf( file, "-" );
             fprintf( file, "%s", list[i]?list[i]:"???" );
